🚚 [RUMF-828] move configuration-related modules in a sub directory
