+++ packages/core/src/domain/configuration/index.ts
@@ -1,0 +1,8 @@
+export { Datacenter } from './transport'
+export {
+  UserConfiguration,
+  buildConfiguration,
+  Configuration,
+  buildCookieOptions,
+  DEFAULT_CONFIGURATION,
+} from './configuration'
