+++ packages/core/src/index.ts
@@ -47,4 +47,4 @@ export { createContextManager } from './tools/contextManager'
 export { limitModification } from './tools/limitModification'
 
 export * from './tools/specHelper'
-export { Datacenter } from './domain/transportConfiguration'
+export { Datacenter } from './domain/configuration'
